Fix : Fix again an ansible issue
It seems we need to remove the ~/.ansible directory if ansible has been already installed before. But we have to remove the directory at the end of the server installation. Because when we will connect by SSH to the host to configure a guest host, ansible won't be able to start.